% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKucoinData.R
\name{GetKucoinData}
\alias{GetKucoinData}
\title{Retrieve criptocurrencies pairs data from Kucoin API}
\usage{
GetKucoinData(criptopair, relevationfrequency, startdate, enddate)
}
\arguments{
\item{criptopair}{tag of the desired cripto pair}

\item{relevationfrequency}{1min, 3min, 5min, 15min, 30min, 1hour, 2hour, 4hour, 6hour, 8hour, 12hour, 1day, 1week}

\item{startdate}{YYYY-MM-DD be careful, a lot of criptopairs are very young}

\item{enddate}{YYYY-MM-DD}
}
\value{
up to 1500 candlestick relevation (very few, must expand this constraint somehow, however that's the first version)
        each relevation is composed by open, close, high, low.
}
\description{
Retrieve criptocurrencies pairs data from Kucoin API
}
\examples{
 W = GetKucoinData("ADA-BTC", "1min", "2019-12-31", "2020-01-31")


}
