% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetonumber.R
\name{getnumericdata}
\alias{getnumericdata}
\title{auxiliary function to allow getkucoindata to work}
\usage{
getnumericdata(date)
}
\arguments{
\item{date}{to translate in number}
}
\value{

}
\description{
auxiliary function to allow getkucoindata to work
}
